refactor: centralize delay and pagination helpers
